+++ src/client/widgets/pads/xy.js
@@ -104,7 +104,7 @@ module.exports = class Xy extends Pad {
             this.faders.x.trigger('dragend', {...e, stopPropagation: true})
             this.faders.y.trigger('dragend', {...e, stopPropagation: true})
             if (this.getProp('spring')) {
-                this.setValue([this.faders.x.getSpringValue(),this.faders.y.getSpringValue()],{sync:true,send:true})
+                this.setValue([this.faders.x.getSpringValue(),this.faders.y.getSpringValue()],{sync:true, send:true, spring:true})
             } else {
                 this.batchDraw()
             }
@@ -123,7 +123,7 @@ module.exports = class Xy extends Pad {
                 doubletab(this.widget, ()=>{
                     this.faders.x.setValue(this.faders.x.getSpringValue(), {sync: false, send:false, dragged:true})
                     this.faders.y.setValue(this.faders.y.getSpringValue(), {sync: false, send:false, dragged:true})
-                    this.setValue([this.faders.x.getSpringValue(),this.faders.y.getSpringValue()],{sync:true, send:true})
+                    this.setValue([this.faders.x.getSpringValue(),this.faders.y.getSpringValue()],{sync:true, send:true, spring:true})
                 })
 
             }