+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - bug fixes
   - script: issue when using the `options` argument in `set()` (options leaked to subsequent set() calls in the script)
    - multixy: `decimals` property not applied
+   - multixy: spring behavior not working until all points are released  
 
 # 1.8.8